+++ packages/sns/src/index.ts
@@ -4,14 +4,10 @@ export {
   type ITopicBuilder,
 } from "./topic-builder.js";
 export { TOPIC_DEFAULTS } from "./defaults.js";
-export { type TopicAlarmConfig } from "./alarm-config.js";
-export { TOPIC_ALARM_DEFAULTS } from "./alarm-defaults.js";
+export { type TopicAlarmConfig } from "./topic-alarm-config.js";
+export { TOPIC_ALARM_DEFAULTS } from "./topic-alarm-defaults.js";
 export {
   createSubscriptionBuilder,
   type ISubscriptionBuilder,
-  type SubscriptionBuilderProps,
   type SubscriptionBuilderResult,
 } from "./subscription-builder.js";
-export { SUBSCRIPTION_DEFAULTS } from "./subscription-defaults.js";
-export { type SubscriptionAlarmConfig } from "./subscription-alarm-config.js";
-export { SUBSCRIPTION_ALARM_DEFAULTS } from "./subscription-alarm-defaults.js";
