+++ packages/sns/src/topic-alarms.ts
@@ -4,8 +4,8 @@ import type { ITopic } from "aws-cdk-lib/aws-sns";
 import type { IConstruct } from "constructs";
 import type { AlarmDefinition } from "@composurecdk/cloudwatch";
 import { AlarmDefinitionBuilder, createAlarms, resolveAlarmConfig } from "@composurecdk/cloudwatch";
-import type { TopicAlarmConfig } from "./alarm-config.js";
-import { TOPIC_ALARM_DEFAULTS } from "./alarm-defaults.js";
+import type { TopicAlarmConfig } from "./topic-alarm-config.js";
+import { TOPIC_ALARM_DEFAULTS } from "./topic-alarm-defaults.js";
 
 const METRIC_PERIOD = Duration.minutes(1);
 const METRIC_PERIOD_LABEL = `${String(METRIC_PERIOD.toMinutes())} minute`;
@@ -55,6 +55,46 @@ export function resolveTopicAlarmDefinitions(
       datapointsToAlarm: cfg.datapointsToAlarm,
       treatMissingData: cfg.treatMissingData,
       description: `SNS topic messages are being filtered out due to invalid subscription filter policy attributes. Threshold: > ${String(cfg.threshold)} filtered messages in ${METRIC_PERIOD_LABEL}.`,
+    });
+  }
+
+  if (config?.numberOfNotificationsRedrivenToDlq !== false) {
+    const cfg = resolveAlarmConfig(
+      config?.numberOfNotificationsRedrivenToDlq,
+      TOPIC_ALARM_DEFAULTS.numberOfNotificationsRedrivenToDlq,
+    );
+    definitions.push({
+      key: "numberOfNotificationsRedrivenToDlq",
+      metric: topic.metric("NumberOfNotificationsRedrivenToDlq", {
+        period: METRIC_PERIOD,
+        statistic: "Sum",
+      }),
+      threshold: cfg.threshold,
+      comparisonOperator: ComparisonOperator.GREATER_THAN_THRESHOLD,
+      evaluationPeriods: cfg.evaluationPeriods,
+      datapointsToAlarm: cfg.datapointsToAlarm,
+      treatMissingData: cfg.treatMissingData,
+      description: `SNS topic is redriving messages to a subscription dead-letter queue, indicating delivery failures. Threshold: > ${String(cfg.threshold)} redrives in ${METRIC_PERIOD_LABEL}.`,
+    });
+  }
+
+  if (config?.numberOfNotificationsFailedToRedriveToDlq !== false) {
+    const cfg = resolveAlarmConfig(
+      config?.numberOfNotificationsFailedToRedriveToDlq,
+      TOPIC_ALARM_DEFAULTS.numberOfNotificationsFailedToRedriveToDlq,
+    );
+    definitions.push({
+      key: "numberOfNotificationsFailedToRedriveToDlq",
+      metric: topic.metric("NumberOfNotificationsFailedToRedriveToDlq", {
+        period: METRIC_PERIOD,
+        statistic: "Sum",
+      }),
+      threshold: cfg.threshold,
+      comparisonOperator: ComparisonOperator.GREATER_THAN_THRESHOLD,
+      evaluationPeriods: cfg.evaluationPeriods,
+      datapointsToAlarm: cfg.datapointsToAlarm,
+      treatMissingData: cfg.treatMissingData,
+      description: `SNS topic failed to redrive a message to a subscription dead-letter queue; messages may be lost. Threshold: > ${String(cfg.threshold)} failed redrives in ${METRIC_PERIOD_LABEL}.`,
     });
   }
